Refactor block registration to handle components + forcing
